Исправление настроек liquibase (changeLogFile, spring.liquibase.change-log). Добавлние spring.jpa.hibernate.ddl-auto=validate и исправление найденных ошибок. Добавлtние скриптов для дефолтных брендов, категорий и продуктов.

+++ shop-database/src/main/resources/db/changelog/v-1.0/01-create-user-and-role-tables.sql
@@ -1,6 +1,6 @@
 create table role
 (
-    id   bigint       not null auto_increment,
+    id   int       not null auto_increment,
     name varchar(255) not null,
     primary key (id)
 ) engine = InnoDB;
@@ -8,7 +8,7 @@ GO
 
 create table user
 (
-    id       bigint       not null auto_increment,
+    id       int       not null auto_increment,
     name     varchar(32)  not null,
     password varchar(128) not null,
     age      integer,
@@ -19,8 +19,8 @@ GO
 
 create table user_role
 (
-    user_id bigint not null,
-    role_id bigint not null,
+    user_id int not null,
+    role_id int not null,
     primary key (user_id, role_id)
 ) engine = InnoDB;
 GO
